pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trending-hunter"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
th = "trending_hunter.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
